rename platform.js to platformUtils.js
