+++ bin/tldr
@@ -4,7 +4,7 @@ const program = require('commander');
 const pkg = require('../package');
 const Tldr = require('../lib/tldr');
 const config = require('../lib/config');
-const platform = require('../lib/platform');
+const platformUtils = require('../lib/platformUtils');
 
 program
   .version(pkg.version, '-v, --version')
@@ -20,11 +20,11 @@ program
   .option('-e, --random-example', 'Show a random example')
   .option('-f, --render [file]', 'Render a specific markdown [file]')
   .option('-m, --markdown', 'Output in markdown format')
-  .option('-p, --platform [type]', 'Override the operating system [linux, osx, sunos, windows]')
-  .option('--linux', 'Override the operating system with Linux')
-  .option('--osx', 'Override the operating system with OSX')
-  .option('--sunos', 'Override the operating system with SunOS')
-  .option('--windows', 'Override the operating system with Windows')
+  .option('-p, --platform [type]', 'Override platform [linux, osx, sunos, windows]')
+  .option('--linux', 'Override platform with Linux')
+  .option('--osx', 'Override platform with OSX')
+  .option('--sunos', 'Override platform with SunOS')
+  .option('--windows', 'Override platform with Windows')
   .option('-t, --theme [theme]', 'Color theme (simple, base16, ocean)')
   .option('-s, --search [keywords]', 'Search pages using keywords')
   //
@@ -61,25 +61,25 @@ program.on('--help', () => {
 program.parse(process.argv);
 
 if (program.linux) {
-  program.os = 'linux';
+  program.platform = 'linux';
 }
 
 if (program.osx) {
-  program.os = 'osx';
+  program.platform = 'osx';
 }
 
 if (program.sunos) {
-  program.os = 'sunos';
+  program.platform = 'sunos';
 }
 
 if (program.windows) {
-  program.os = 'windows';
+  program.platform = 'windows';
 }
 
 let cfg = config.get();
-if (program.os) {
-  if (platform.isSupported(program.os)) {
-    cfg.platform = program.os;
+if (program.platform) {
+  if (platformUtils.isSupported(program.platform)) {
+    cfg.platform = program.platform;
   }
 }
