+++ lib/cache.js
@@ -4,7 +4,7 @@ const fs = require('fs-extra');
 const os = require('os');
 const path = require('path');
 const remote = require('./remote');
-const platform = require('./platform');
+const platformUtils = require('./platformUtils');
 const index = require('./index');
 const utils = require('./utils');
 
@@ -21,7 +21,7 @@ class Cache {
   }
 
   getPage(page) {
-    let preferredPlatform = platform.getPreferredPlatformFolder(this.config);
+    let preferredPlatform = platformUtils.getPreferredPlatformFolder(this.config);
     const preferredLanguage = process.env.LANG || 'en';
     return index.findPage(page, preferredPlatform, preferredLanguage)
       .then((folder) => {
