+++ lib/config.js
@@ -4,6 +4,7 @@ const defaults = require('lodash/defaults');
 const fs = require('fs');
 const path = require('path');
 const osHomedir = require('os').homedir;
+let platformUtils = require('./platformUtils');
 
 exports.get = () => {
   const DEFAULT = path.join(__dirname, '..', 'config.json');
@@ -38,8 +39,7 @@ exports.get = () => {
 };
 
 function validatePlatform(os) {
-  let platform = require('./platform');
-  if (os && !platform.isSupported(os)) {
+  if (os && !platformUtils.isSupported(os)) {
     return 'Unsupported platform : ' + os;
   }
   return null;
