+++ lib/platformUtils.js
