+++ lib/search.js
@@ -7,7 +7,7 @@ const natural = require('natural');
 const config = require('./config');
 const utils = require('./utils');
 const index = require('./index');
-const platform = require('./platform');
+const platformUtils = require('./platformUtils');
 
 const CACHE_FOLDER = path.join(config.get().cache, 'cache');
 
@@ -193,7 +193,7 @@ exports.printResults = (results, config) => {
   //              $ tree (Available on: linux, osx)
   index.getShortIndex().then((shortIndex) => {
     let outputs = new Set();
-    let preferredPlatform = platform.getPreferredPlatform(config);
+    let preferredPlatform = platformUtils.getPreferredPlatform(config);
     results.forEach((elem) => {
       let cmdname = utils.parsePagename(elem.file);
       let output = '    $ ' + cmdname;
