+++ lib/tldr.js
@@ -6,7 +6,7 @@ const ms = require('ms');
 const ora = require('ora');
 const Cache = require('./cache');
 const search = require('./search');
-const platform = require('./platform');
+const platformUtils = require('./platformUtils');
 const messages = require('./messages');
 const parser = require('./parser');
 const render = require('./render');
@@ -22,7 +22,7 @@ class Tldr {
   }
 
   list(singleColumn) {
-    let os = platform.getPreferredPlatformFolder(this.config);
+    let os = platformUtils.getPreferredPlatformFolder(this.config);
     return index.commandsFor(os)
       .then((commands) => {
         return this.printPages(commands, singleColumn);
@@ -41,7 +41,7 @@ class Tldr {
   }
 
   random(options) {
-    let os = platform.getPreferredPlatformFolder(this.config);
+    let os = platformUtils.getPreferredPlatformFolder(this.config);
     return index.commandsFor(os)
       .then((pages) => {
         if (pages.length === 0) {
@@ -57,7 +57,7 @@ class Tldr {
   }
 
   randomExample() {
-    let os = platform.getPreferredPlatformFolder(this.config);
+    let os = platformUtils.getPreferredPlatformFolder(this.config);
     return index.commandsFor(os)
       .then((pages) => {
         if (pages.length === 0) {
