+++ test/platform.spec.js
@@ -3,7 +3,7 @@
 const os = require('os');
 const config = require('../lib/config');
 const sinon = require('sinon');
-const platform = require('../lib/platform');
+const platformUtils = require('../lib/platformUtils');
 
 describe('Platform', () => {
 
@@ -20,7 +20,7 @@ describe('Platform', () => {
     it('should return the running platform with no configuration', () => {
       os.platform.onCall(0).returns('darwin');
       this.config = {};
-      platform.getPreferredPlatform(this.config).should.eql('darwin');
+      platformUtils.getPreferredPlatform(this.config).should.eql('darwin');
     });
 
     it('should overwrite the running platform if configured', () => {
@@ -28,7 +28,7 @@ describe('Platform', () => {
       this.config = {
         platform: 'linux'
       };
-      platform.getPreferredPlatform(this.config).should.eql('linux');
+      platformUtils.getPreferredPlatform(this.config).should.eql('linux');
     });
 
     it('should return current system platform if configuration is wrong', () => {
@@ -36,17 +36,17 @@ describe('Platform', () => {
       this.config = {
         platform: 'there_is_no_such_platform'
       };
-      platform.getPreferredPlatform(this.config).should.eql('darwin');
+      platformUtils.getPreferredPlatform(this.config).should.eql('darwin');
     });
   });
 
   describe('isSupported', () => {
     it('should tell that Linux, OSX, SunOS and Win32 are supported', () => {
-      platform.isSupported('osx').should.eql(true);
-      platform.isSupported('linux').should.eql(true);
-      platform.isSupported('sunos').should.eql(true);
-      platform.isSupported('windows').should.eql(true);
-      platform.isSupported('ios').should.eql(false);
+      platformUtils.isSupported('osx').should.eql(true);
+      platformUtils.isSupported('linux').should.eql(true);
+      platformUtils.isSupported('sunos').should.eql(true);
+      platformUtils.isSupported('windows').should.eql(true);
+      platformUtils.isSupported('ios').should.eql(false);
     });
   });
 });
